Move batch to utility
